Külön komponensbe lett átmozgatva a kártya
